+++ plugins/org.ietr.preesm.mapper/src/org/ietr/preesm/mapper/graphtransfo/SdfToDagConverter.java
@@ -48,15 +48,11 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.logging.Level;
-import org.ietr.dftools.algorithm.model.AbstractEdge;
-import org.ietr.dftools.algorithm.model.AbstractEdgePropertyType;
 import org.ietr.dftools.algorithm.model.dag.DAGEdge;
 import org.ietr.dftools.algorithm.model.dag.DAGVertex;
-import org.ietr.dftools.algorithm.model.dag.types.DAGDefaultEdgePropertyType;
-import org.ietr.dftools.algorithm.model.parameters.InvalidExpressionException;
-import org.ietr.dftools.algorithm.model.sdf.SDFAbstractVertex;
 import org.ietr.dftools.algorithm.model.sdf.SDFEdge;
 import org.ietr.dftools.algorithm.model.sdf.SDFGraph;
+import org.ietr.dftools.algorithm.model.sdf.types.SDFIntEdgePropertyType;
 import org.ietr.dftools.algorithm.model.sdf.visitors.DAGTransformation;
 import org.ietr.dftools.algorithm.model.visitors.SDF4JException;
 import org.ietr.dftools.architecture.slam.ComponentInstance;
@@ -113,6 +109,7 @@ public class SdfToDagConverter {
       return null;
     }
     final SDFGraph sdf = sdfIn.clone();
+    setDataSizeForSDF(sdf, scenario);
     // Generates a dag
     final MapperDAG dag = new MapperDAG(new MapperEdgeFactory(), sdf);
 
@@ -135,8 +132,7 @@ public class SdfToDagConverter {
       WorkflowLogger.getLogger().log(Level.SEVERE, "Can not map a DAG with no vertex.");
     } else {
       WorkflowLogger.getLogger().log(Level.INFO, "Conversion finished.");
-      WorkflowLogger.getLogger().log(Level.INFO,
-          "mapping a DAG with " + dag.vertexSet().size() + " vertices and " + dag.edgeSet().size() + " edges.");
+      WorkflowLogger.getLogger().log(Level.INFO, "mapping a DAG with " + dag.vertexSet().size() + " vertices and " + dag.edgeSet().size() + " edges.");
     }
 
     scenario.getDAGs2SDFs().put(dag.getName(), sdfIn);
@@ -155,8 +151,7 @@ public class SdfToDagConverter {
    *          the scenario
    * @return The DAG with initial properties
    */
-  public static MapperDAG addInitialProperties(final MapperDAG dag, final Design architecture,
-      final PreesmScenario scenario) {
+  public static MapperDAG addInitialProperties(final MapperDAG dag, final Design architecture, final PreesmScenario scenario) {
 
     SdfToDagConverter.addInitialVertexProperties(dag, architecture, scenario);
     SdfToDagConverter.addInitialEdgeProperties(dag, architecture, scenario);
@@ -165,6 +160,22 @@ public class SdfToDagConverter {
     SdfToDagConverter.addInitialRelativeConstraintsProperties(dag, architecture, scenario);
     SdfToDagConverter.addInitialTimingProperties(dag, architecture, scenario);
     return null;
+  }
+
+  /**
+   * Sets the data size property of every edge of the HSDF graph
+   *
+   * @param graph
+   *          the graph
+   * @param scenario
+   *          the scenario
+   */
+  private static void setDataSizeForSDF(final SDFGraph graph, final PreesmScenario scenario) {
+    for (final SDFEdge edge : graph.edgeSet()) {
+      final String type = edge.getDataType().toString();
+      final int size = scenario.getSimulationManager().getDataTypeSizeOrDefault(type);
+      edge.setDataSize(new SDFIntEdgePropertyType(size));
+    }
   }
 
   /**
@@ -178,8 +189,7 @@ public class SdfToDagConverter {
    * @param scenario
    *          the scenario
    */
-  public static void addInitialRelativeConstraintsProperties(final MapperDAG dag, final Design architecture,
-      final PreesmScenario scenario) {
+  public static void addInitialRelativeConstraintsProperties(final MapperDAG dag, final Design architecture, final PreesmScenario scenario) {
 
     // Initial relative constraints are stored in the scenario
     final RelativeConstraintManager manager = scenario.getRelativeconstraintManager();
@@ -223,8 +233,7 @@ public class SdfToDagConverter {
    * @param scenario
    *          the scenario
    */
-  public static void addInitialTimingProperties(final MapperDAG dag, final Design architecture,
-      final PreesmScenario scenario) {
+  public static void addInitialTimingProperties(final MapperDAG dag, final Design architecture, final PreesmScenario scenario) {
 
     // New timing objects are created they are not initialized here
     for (final DAGVertex v : dag.vertexSet()) {
@@ -283,8 +292,7 @@ public class SdfToDagConverter {
    * @param scenario
    *          the scenario
    */
-  public static void addInitialVertexProperties(final MapperDAG dag, final Design architecture,
-      final PreesmScenario scenario) {
+  public static void addInitialVertexProperties(final MapperDAG dag, final Design architecture, final PreesmScenario scenario) {
 
     /**
      * Importing default timings
@@ -301,8 +309,7 @@ public class SdfToDagConverter {
       currentVertexInit.setNbRepeat(nbRepeat);
 
       // The SDF vertex id is used to reference the timings
-      final List<Timing> timelist = scenario.getTimingManager().getGraphTimings(currentVertex,
-          DesignTools.getOperatorComponentIds(architecture));
+      final List<Timing> timelist = scenario.getTimingManager().getGraphTimings(currentVertex, DesignTools.getOperatorComponentIds(architecture));
 
       // Iterating over timings for each DAG vertex
       final Iterator<Timing> listiterator = timelist.iterator();
@@ -322,8 +329,7 @@ public class SdfToDagConverter {
         } else {
           // Default timings are given
           for (final ComponentInstance op : DesignTools.getOperatorInstances(architecture)) {
-            final Timing time = new Timing(op.getComponent().getVlnv().getName(), currentVertex.getId(),
-                Timing.DEFAULT_TASK_TIME);
+            final Timing time = new Timing(op.getComponent().getVlnv().getName(), currentVertex.getId(), Timing.DEFAULT_TASK_TIME);
             currentVertexInit.addTiming(time);
           }
         }
@@ -341,8 +347,7 @@ public class SdfToDagConverter {
    * @param scenario
    *          the scenario
    */
-  public static void addInitialSpecialVertexProperties(final MapperDAG dag, final Design architecture,
-      final PreesmScenario scenario) {
+  public static void addInitialSpecialVertexProperties(final MapperDAG dag, final Design architecture, final PreesmScenario scenario) {
 
     // Iterating over dag vertices
     final TopologicalDAGIterator dagiterator = new TopologicalDAGIterator(dag);
@@ -360,8 +365,7 @@ public class SdfToDagConverter {
           final Timing timing = new Timing(opDef, currentVertex.getId());
 
           // Depending on the type of vertex, time is given by the size of output or input buffers
-          if (SpecialVertexManager.isFork(currentVertex) || SpecialVertexManager.isJoin(currentVertex)
-              || SpecialVertexManager.isEnd(currentVertex)) {
+          if (SpecialVertexManager.isFork(currentVertex) || SpecialVertexManager.isJoin(currentVertex) || SpecialVertexManager.isEnd(currentVertex)) {
             timing.setTime(sut + (long) (tpu * SdfToDagConverter.getVertexInputBuffersSize(currentVertex)));
           } else if (SpecialVertexManager.isBroadCast(currentVertex) || SpecialVertexManager.isInit(currentVertex)) {
             timing.setTime(sut + (long) (tpu * SdfToDagConverter.getVertexOutputBuffersSize(currentVertex)));
@@ -384,8 +388,7 @@ public class SdfToDagConverter {
    *          the scenario
    */
   @SuppressWarnings("unchecked")
-  public static void addInitialEdgeProperties(final MapperDAG dag, final Design architecture,
-      final PreesmScenario scenario) {
+  public static void addInitialEdgeProperties(final MapperDAG dag, final Design architecture, final PreesmScenario scenario) {
 
     /**
      * Importing data edge weights and multiplying by type size when available
@@ -395,37 +398,7 @@ public class SdfToDagConverter {
     while (edgeiterator.hasNext()) {
       final MapperDAGEdge currentEdge = (MapperDAGEdge) edgeiterator.next();
       final EdgeInit currentEdgeInit = currentEdge.getInit();
-
-      int weight = 0;
-      if (currentEdge.getAggregate().isEmpty()) {
-        final DAGDefaultEdgePropertyType dagDefaultEdgePropertyType = (DAGDefaultEdgePropertyType) currentEdge
-            .getWeight();
-        weight = dagDefaultEdgePropertyType.intValue();
-        final String dataType = currentEdge.getPropertyBean()
-            .getValue(SDFEdge.DATA_TYPE, AbstractEdgePropertyType.class).toString();
-        final int typeSize = scenario.getSimulationManager().getDataTypeSizeOrDefault(dataType);
-        weight *= typeSize;
-      } else {
-
-        // Calculating the edge weight for simulation:
-        // edge production * number of source repetitions * type size
-        for (final AbstractEdge<SDFGraph, SDFAbstractVertex> aggMember : currentEdge.getAggregate()) {
-          final SDFEdge sdfAggMember = (SDFEdge) aggMember;
-          int prod;
-          try {
-            prod = sdfAggMember.getProd().intValue();
-          } catch (final InvalidExpressionException e) {
-            // TODO Auto-generated catch block
-            e.printStackTrace();
-            prod = 1;
-          }
-          final int nbRepeat = currentEdge.getSource().getNbRepeat().intValue();
-          final int typeSize = scenario.getSimulationManager()
-              .getDataTypeSizeOrDefault(sdfAggMember.getDataType().toString());
-          weight += prod * nbRepeat * typeSize;
-        }
-      }
-      currentEdgeInit.setDataSize(weight);
+      currentEdgeInit.setDataSize(currentEdge.getWeight().intValue());
 
     }
   }
@@ -440,10 +413,9 @@ public class SdfToDagConverter {
    * @param scenario
    *          the scenario
    */
-  public static void addInitialConstraintsProperties(final MapperDAG dag, final Design architecture,
-      final PreesmScenario scenario) {
+  public static void addInitialConstraintsProperties(final MapperDAG dag, final Design architecture, final PreesmScenario scenario) {
     /**
-     * Importing scenario: Only the timings corresponding to allowed mappings are set.
+     * Importing scenario: Only the timings to allowed mappings are set.
      */
 
     // Iterating over constraint groups