+++ release_notes.md
@@ -34,6 +34,7 @@ PREESM Changelog
 * Have travis retry 3 times to fetch dependencies;
 * Force DFTools to 1.4.1+;
 * Enforced use of DAG properties instead of SDF properties when conversion of an SDF graph is done
+* Initialization of data size property in SDFEdge before dag conversion. 
 
 ### Bug fix
 * Fix codegen for MS Visual Studio to support thread affinity;